requirements.txt
streamlit
openai
langchain
llama-index
transformers
torch
psycopg2
python-dotenv
fuzzywuzzy